split functionality by database engine type
The plugin as it stands today is increasing complexity of queries given different engine types
Current Engine type ssupported Azure SQL DB, Managed instance, SQL Server versions 2008 - 2019
Config file now has a database_type which is SQLServer / AzureSQLDB / AzureSQLManagedInstance.
This enables us to split up the queries so one change doesn't regress others and reduce testing surface.
This will take only if users use database_type in config, if not it remains as before.
Also split queries for V1 and queries for Azure SQL into separate files.
